fix clippy issues

+++ projects/ledger_sqn/src/lib.rs
@@ -21,6 +21,6 @@ pub extern "C" fn finish() -> i32 {
         .unwrap();
 
         let _ = trace_num("Ledger Sequence", ledger_sequence as i64);
-        ledger_sequence >= 5
+        (ledger_sequence >= 5) as i32 // Return 1 if true, 0 if false
     }
 }
